+++ src/main/java/dto/SongList.java
@@ -1,3 +1,5 @@
+package dto;
+
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
