+++ src/main/java/service/StartMerging.java
@@ -1,3 +1,11 @@
+package service;
+
+import dto.SongList;
+import filter.Filter;
+import merger.Merger;
+import parser.ExcelParser;
+import tools.ToplijstenMergerTools;
+
 import java.util.*;
 
 public class StartMerging {
