+++ src/test/java/dto/TestSong.java
@@ -1,7 +1,9 @@
+package dto;
+
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 
-class TestRecord {
+class TestSong {
 
     @Test
     void testRecord() {
