+++ src/test/java/parser/TestSpecialCharacters.java
@@ -1,9 +1,29 @@
-import java.io.File;
+package parser;
 
+import dto.Song;
+import dto.SongList;
 import org.junit.jupiter.api.Test;
+import util.FileUtil;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-class TestExcelParser {
+import java.io.File;
+
+class TestSpecialCharacters {
+
+    @Test
+    void testReadAmpersands() {
+        final XMLParser parser = new XMLParser();
+        SongList songList = new SongList();
+        final File inFile = FileUtil.getFileFromResourceRootAndOtherwiseFromCustomPath("testampersands.xml");
+        final String fileToFeed = inFile.toString();
+        songList = parser.parseXML(fileToFeed, songList);
+
+        final Song firstSong = songList.get(0);
+        assertThat(firstSong.getTitle()).isEqualTo("Music &amp; Lights");
+        final Song secondSong = songList.get(1);
+        assertThat(secondSong.getTitle()).isEqualTo("Music&amp;Lights");
+    }
 
     @Test
     void testRecord() {
