+++ src/test/java/util/FileUtil.java
@@ -1,3 +1,5 @@
+package util;
+
 import java.io.File;
 
 public class FileUtil {
